+++ src/Select.svelte
@@ -26,6 +26,16 @@
     export let items = null;
     export let itemFilter = (label, filterText, option) =>
         `${label}`.toLowerCase().includes(filterText.toLowerCase());
+    export let searchScore = (label, filterText, item) => {
+        if (filterText === '') return Number.MAX_VALUE;
+        label = label.toLowerCase().trim();
+        filterText = filterText.toLowerCase().trim();
+        let matches = 0;
+        for (let i = 0; i < label.length && i < filterText.length; ++i) {
+            if (label[i] === filterText[i]) matches++;
+        }
+        return matches;
+    };
     export let groupBy = undefined;
     export let groupFilter = (groups) => groups;
     export let isGroupHeaderSelectable = false;
@@ -65,8 +75,10 @@
         return filteredItems;
     };
 
-    export let isSearchable; // Deprecated. Use isFilterable instead.
     export let isFilterable = true;
+    export let isSearchable = false;
+    export let searchResults = 5;
+    export let minSearchScore = Number.MIN_VALUE;
     export let inputStyles = '';
     export let isClearable = true;
     export let isWaiting = false;
@@ -95,11 +107,8 @@
     export let VirtualList = _VirtualList;
 
     $: {
-        if (isSearchable !== undefined) {
-            isFilterable = isSearchable;
-            console.warn(
-                'isSearchable is deprecated and will change in the future. Use isFilterable instead.'
-            );
+        if (isSearchable) {
+            isFilterable = false;
         }
     }
 
@@ -115,28 +124,57 @@
             args.items = convertStringItemsToObjects(args.items);
         }
 
-        let filterResults = args.items.filter((item) => {
-            let matchesFilter = itemFilter(
-                getOptionLabel(item, args.filterText),
-                args.filterText,
-                item
-            );
+        let filterResults;
 
-            if (
-                matchesFilter &&
-                args.isMulti &&
-                args.value &&
-                Array.isArray(args.value)
-            ) {
-                matchesFilter = !args.value.some((x) => {
-                    return (
-                        x[args.optionIdentifier] === item[args.optionIdentifier]
-                    );
-                });
+        if (isFilterable) {
+            filterResults = args.items.filter((item) => {
+                let matchesFilter = itemFilter(
+                    getOptionLabel(item, args.filterText),
+                    args.filterText,
+                    item
+                );
+
+                if (
+                    matchesFilter &&
+                    args.isMulti &&
+                    args.value &&
+                    Array.isArray(args.value)
+                ) {
+                    matchesFilter = !args.value.some((x) => {
+                        return (
+                            x[args.optionIdentifier] ===
+                            item[args.optionIdentifier]
+                        );
+                    });
+                }
+
+                return matchesFilter;
+            });
+        } else if (isSearchable) {
+            filterResults = [];
+            for (let i = 0; i < args.items.length; ++i) {
+                let item = args.items[i];
+                item._score = searchScore(
+                    getOptionLabel(item, args.filterText),
+                    args.filterText,
+                    item
+                );
+                if (item._score >= minSearchScore) {
+                    filterResults.push(item);
+                }
             }
 
-            return matchesFilter;
-        });
+            filterResults = filterResults.sort((left, right) => {
+                return right._score - left._score;
+            });
+
+            if (searchResults > 0) {
+                filterResults = filterResults.slice(0, searchResults);
+            }
+            filterResults.forEach((item) => {
+                delete item._score;
+            });
+        }
 
         if (args.groupBy) {
             filterResults = filterGroupedItems(filterResults);
@@ -246,7 +284,7 @@
             _inputAttributes.id = id;
         }
 
-        if (!isFilterable) {
+        if (!(isFilterable || isSearchable)) {
             _inputAttributes.readonly = true;
         }
     }
@@ -343,7 +381,8 @@
     }
 
     $: {
-        if (inputAttributes || !isFilterable) assignInputAttributes();
+        if (inputAttributes || !(isFilterable || isSearchable))
+            assignInputAttributes();
     }
 
     $: {
@@ -883,9 +922,7 @@
         class="a11yText">
         {#if isFocused}
             <span id="aria-selection">{ariaSelection}</span>
-            <span id="aria-context">
-                {ariaContext}
-            </span>
+            <span id="aria-context"> {ariaContext} </span>
         {/if}
     </span>
 
@@ -906,7 +943,7 @@
     {/if}
 
     <input
-        readOnly={!isFilterable}
+        readOnly={!(isFilterable || isSearchable)}
         {..._inputAttributes}
         bind:this={input}
         on:focus={handleFocus}
@@ -933,7 +970,7 @@
         </div>
     {/if}
 
-    {#if !showClearIcon && (showIndicator || (showChevron && !value) || (!isFilterable && !isDisabled && !isWaiting && ((showSelectedItem && !isClearable) || !showSelectedItem)))}
+    {#if !showClearIcon && (showIndicator || (showChevron && !value) || (!(isFilterable || isSearchable) && !isDisabled && !isWaiting && ((showSelectedItem && !isClearable) || !showSelectedItem)))}
         <div class="indicator" aria-hidden="true">
             {#if indicatorSvg}
                 {@html indicatorSvg}