+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # svelte-select changelog
 
-## 4.5.0
+## <next>
 
-* Renamed option `isSearchable` to `isFilterable`. `isSearchable` is now deprecated
-    and will have different behaviour in the future.
+* Added option `isSearchable` that enables search functionality as opposed to filtering.
+* Added method `searchScore` that calculates a search score for a given item against the filterText.
+* Added option `searchResults` that restricts search results to the top n search results.
+    Defaults to `5` when searching is enabled.
+* Added option `minSearchScore` that puts a lower search boundary which hides results below that score.
 
 ## 4.4.3